Add files via upload
1)When extracting 'Starting 11' events from event list, the result was not following the match_id order, and not any chronological order (like match_game or date)
As an exaple of the error:
In team.matches(for w==1), match_id n. 19758 happens only after match_id 19806.
But in events.list, match_id 19758 is after 19757. This list is ordered by id and not by date(or match_week). Thus, the error
This was fixed by imposing the order of the results to the order selected previously.
  
2)Added dependencies: dplyr, ggplot

3)%!in% is now defined
 
4)The match week does not follow chronological order. So, it is best to sort by match date